brush up files

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -4,11 +4,8 @@ import net.engineeringdigest.journalApp.entity.JournalEntry;
 import net.engineeringdigest.journalApp.entity.User;
 import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -19,28 +16,26 @@ public class JournalEntryService {
 
     @Autowired
     private JournalEntryRepository journalEntryRepository;
+    
     @Autowired
     private UserService userService;
-
-    private static final Logger logger = LoggerFactory.getLogger(JournalEntryService.class);
 
     public List<JournalEntry> getAll() {
         return journalEntryRepository.findAll();
     }
 
-    @Transactional
     public JournalEntry saveNewEntry(JournalEntry journalEntry, String userName) {
-        try{
-            journalEntry.setDate(LocalDateTime.now());
-            User userInDb = userService.findByUserName(userName);
-            JournalEntry saved = journalEntryRepository.save(journalEntry);
-            userInDb.getJournalEntries().add(saved);
-            userService.saveUser(userInDb);
-            return saved;
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+        journalEntry.setDate(LocalDateTime.now());
+        User userInDb = userService.findByUserName(userName);
+        if (userInDb == null) {
+            throw new RuntimeException("User not found");
         }
+        JournalEntry saved = journalEntryRepository.save(journalEntry);
+        userInDb.getJournalEntries().add(saved);
+        userService.saveUser(userInDb);
+        return saved;
     }
+
     public void saveEntry(JournalEntry journalEntry) {
         journalEntryRepository.save(journalEntry);
     }
@@ -49,28 +44,16 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    @Transactional
     public boolean deleteById(ObjectId id, String userName) {
-        boolean removed = false;
-        try {
-            User userInDb = userService.findByUserName(userName);
-            removed = userInDb.getJournalEntries().removeIf(x -> x.getId().equals(id));
-            if (removed) {
-                userService.saveUser(userInDb);
-                journalEntryRepository.deleteById(id);
-            }
-        }catch (Exception e) {
-            throw new RuntimeException("An error occurred while deleting the entry.",e);
+        User userInDb = userService.findByUserName(userName);
+        if (userInDb == null) {
+            return false;
+        }
+        boolean removed = userInDb.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if (removed) {
+            userService.saveUser(userInDb);
+            journalEntryRepository.deleteById(id);
         }
         return removed;
     }
-
-    public JournalEntry updateEntry(ObjectId id, JournalEntry journalEntry) {
-        journalEntry.setId(id);
-        return journalEntryRepository.save(journalEntry);
-    }
-
-//    public List<JournalEntry> findByUserName(String userName) {
-//
-//    }
 }
